+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `compact_mode` will be deprecated in version 1.0.0.
 
+- JSON_Processor (to get weather in compact mode).
+
 ## [0.3.3] - 2024-03-22
 
 ### Added